refactor : menugroup 리팩토링
