+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroup.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.tobe.domain.menugroup;
 
-import kitchenpos.common.model.DisplayedName;
+import kitchenpos.menus.tobe.vo.DisplayedName;
 
 import javax.persistence.*;
 import java.util.UUID;
